+++ src/main/java/rx/broadcast/BroadcastOrder.java
@@ -4,5 +4,5 @@ import java.util.function.Consumer;
 
 @FunctionalInterface
 public interface BroadcastOrder<T> {
-    void receive(int sender, Consumer<T> consumer, Timestamped<T> message);
+    void receive(long sender, Consumer<T> consumer, Timestamped<T> message);
 }
